+++ devel/sclmisc/sclmisc.cpp
@@ -1081,7 +1081,7 @@ bso::sBool sclmisc::LoadAndTranslateTags(
 qRH
 	TOL_CBUFFER___ Buffer;
 qRB
-	if ( Found = Load( FileName, Registry, Needness, Content ) )
+	if ( ( Found = Load( FileName, Registry, Needness, Content ) ) )
 		scllocale::TranslateTags( Content, sclrgstry::GetLanguage( Registry, Buffer ), Marker );
 qRR
 qRT
@@ -1129,7 +1129,7 @@ qRH;
 qRB;
 	Filename.Init();
 
-	if ( Found = sclrgstry::BGetValue( Registry, FilenameEntry, Needness, Filename ) )
+	if ( ( Found = sclrgstry::BGetValue( Registry, FilenameEntry, Needness, Filename ) ) )
 		LoadXMLAndTranslateTags( Filename, sclrgstry::GetLanguage( Registry, Buffer ), Content, Level, Marker );
 qRR;
 qRT;
@@ -1282,7 +1282,7 @@ namespace {
 		str::string Id;
 	qRB
 		Id.Init();
-		if ( Exists = sclmisc::OGetValue( rgstry::tentry___( sclrgstry::parameter::targeted_plugin::Id, Target ), Id ) )
+		if ( ( Exists = sclmisc::OGetValue( rgstry::tentry___( sclrgstry::parameter::targeted_plugin::Id, Target ), Id ) ) )
 			FillPluginRelatedTags_( Target, Id, Tags );
 	qRR
 	qRT
@@ -1364,7 +1364,7 @@ qRH
 qRB
 	Tags.Init();
 
-	if ( Exists = GetPluginRelatedTags_( Target, Tags ) ) {
+	if ( ( Exists = GetPluginRelatedTags_( Target, Tags ) ) ) {
 		GetPluginFeatures_( Target, Tags, Filename, Configuration, Locale );
 	
 		sclmisc::MGetValue( rgstry::tentry___( sclrgstry::parameter::TargetedPlugin, Target ), Arguments );